fix duplicate symbol
